Updating...

+++ modules/extras/extras.php
@@ -287,34 +287,31 @@ function exec_ogp_module()
 
 	if(isset($_POST['remove']))
 	{
-		$install_nfo = DATA_PATH . str_replace(' ','_',$_POST['folder']) . ".nfo";
-		if( file_exists($install_nfo) )
-		{
-			$lines = file($install_nfo);
-			unset($lines[0]);// timestamp
-			unset($lines[1]);// nfo file
-			usort($lines, "deeperPathFirst");
-			foreach($lines as $file)
-			{
-				$file = trim($file);
-				if(file_exists($file))
-				{
-					unlink($file);
-					$parent_directory = dirname($file);
-					while(count(scandir($parent_directory)) == 2)
-					{							
-						if(realpath($parent_directory) == realpath($baseDir))
-							break;						
-						if(is_writable($parent_directory))
-							rmdir($parent_directory);
-						else
-							break;
-						$parent_directory = dirname($parent_directory);
-					}
+		$remove = $_POST['remove'];
+		$folderToDelete = str_replace(' ','_',$_POST['folder']);
+		
+		if(isset($folderToDelete) && !empty($folderToDelete)){
+		
+			// Delete nfo file if it exists
+			$install_nfo = DATA_PATH . $folderToDelete . ".nfo";
+			if(file_exists($install_nfo)){
+				unlink($install_nfo);
+				
+				// Delete directory if it exists 
+				// We won't run this operation if the nfo file doesn't exist because it could be abused...
+				$dirToDelete = $remove . "/" . strtolower($folderToDelete);
+				if(file_exists($dirToDelete) && is_dir($dirToDelete)){
+					recursiveDelete($dirToDelete);
+				}
+				
+				// In case we are using case sensitive names... which happens for themes
+				$dirToDelete = $remove . "/" . $folderToDelete;
+				if(file_exists($dirToDelete) && is_dir($dirToDelete)){
+					recursiveDelete($dirToDelete);
 				}
 			}
-			unlink($install_nfo);
 		}
+		
 		return;
 	}
